Refactor and relocate connect_to_tor
